P4ADEV-4539 Centralizes ingestion file lock acquisition in base WF
