+++ src/main/java/it/gov/pagopa/pu/workflow/wf/ingestionflow/debtposition/wfingestion/DebtPositionIngestionFlowWFImpl.java
@@ -11,6 +11,7 @@ import it.gov.pagopa.payhub.activities.dto.ingestion.debtposition.SyncIngestedDe
 import it.gov.pagopa.pu.workflow.utilities.Constants;
 import it.gov.pagopa.pu.workflow.utilities.TaskQueueConstants;
 import it.gov.pagopa.pu.workflow.wf.ingestionflow.BaseIngestionFlowFileWFImpl;
+import it.gov.pagopa.pu.workflow.wf.ingestionflow.BaseOrganizationSequentialIngestionFlowFileWFImpl;
 import it.gov.pagopa.pu.workflow.wf.ingestionflow.debtposition.activity.StartMassiveNoticesGenerationWFActivity;
 import it.gov.pagopa.pu.workflow.wf.ingestionflow.debtposition.config.DebtPositionIngestionFlowWfConfig;
 import lombok.extern.slf4j.Slf4j;
@@ -22,14 +23,7 @@ import java.util.function.Function;
 
 @Slf4j
 @WorkflowImpl(taskQueues = TaskQueueConstants.TASK_QUEUE_IMPORT_MEDIUM_PRIORITY)
-public class DebtPositionIngestionFlowWFImpl extends BaseIngestionFlowFileWFImpl<InstallmentIngestionFlowFileResult> implements DebtPositionIngestionFlowWF {
-
-  private static final Duration SLEEP_BETWEEN_ACQUIRE_LOCK = Duration.ofSeconds(5);
-  /**
-   * The lock acquire max attempts before to clear Temporal history.
-   * The threshold is very high ({@link Constants#THRESHOLD_TEMPORAL_EVENTS_BEFORE_CONTINUE_AS_NEW}), lock acquire is the first activity called, we are not interested on WF history, we will clear it before real limit
-   */
-  private static final int LOCK_ATTEMPTS_BEFORE_CLEAN_WF_HISTORY = 1000;
+public class DebtPositionIngestionFlowWFImpl extends BaseOrganizationSequentialIngestionFlowFileWFImpl<InstallmentIngestionFlowFileResult> implements DebtPositionIngestionFlowWF {
 
   private IngestionFlowFileProcessingLockerActivity ingestionFlowFileProcessingLockerActivity;
   private SynchronizeIngestedDebtPositionActivity synchronizeIngestedDebtPositionActivity;
@@ -48,28 +42,6 @@ public class DebtPositionIngestionFlowWFImpl extends BaseIngestionFlowFileWFImpl
   }
 
   @Override
-  protected void setProcessingStatus(Long ingestionFlowFileId) {
-    log.info("Acquiring lock for ingestionFlowFileId {}", ingestionFlowFileId);
-    acquireLock(ingestionFlowFileId);
-    log.info("Lock successfully acquired for ingestionFlowFileId {}", ingestionFlowFileId);
-  }
-
-  private void acquireLock(Long ingestionFlowFileId) {
-    int attemptCounter = 0;
-    while (!ingestionFlowFileProcessingLockerActivity.acquireIngestionFlowFileProcessingLock(ingestionFlowFileId)) {
-      attemptCounter++;
-
-      if (attemptCounter >= LOCK_ATTEMPTS_BEFORE_CLEAN_WF_HISTORY) {
-        log.info("Max attempts reached, continuing as new for ingestionFlowFileId {}", ingestionFlowFileId);
-        Workflow.continueAsNew(ingestionFlowFileId);
-      }
-
-      log.info("Lock not acquired, retrying for ingestionFlowFileId {}", ingestionFlowFileId);
-      Workflow.sleep(SLEEP_BETWEEN_ACQUIRE_LOCK);
-    }
-  }
-
-  @Override
   public void finallyAfterProcessing(Long ingestionFlowFileId, IngestionFlowFileResult ingestionResult) {
     SyncIngestedDebtPositionDTO syncDpResult
       = synchronizeIngestedDebtPositionActivity.synchronizeIngestedDebtPosition(ingestionFlowFileId);
@@ -79,5 +51,10 @@ public class DebtPositionIngestionFlowWFImpl extends BaseIngestionFlowFileWFImpl
     if (StringUtils.isNotBlank(syncDpResult.getPdfGeneratedId())) {
       startMassiveNoticesGenerationWFActivity.startMassiveNoticesGenerationWF(ingestionFlowFileId);
     }
+  }
+
+  @Override
+  protected IngestionFlowFileProcessingLockerActivity getIngestionFlowFileProcessingLockerActivity() {
+    return ingestionFlowFileProcessingLockerActivity;
   }
 }
